Refactor bottom actions
