+++ src/exercises/Exercises.js
@@ -6,12 +6,11 @@ import MultipleChoice from "./exerciseTypes/multipleChoice/MultipleChoice";
 import Congratulations from "./Congratulations";
 import ProgressBar from "./ProgressBar";
 import * as s from "./Exercises.sc";
-import FeedbackButtons from "./FeedbackButtons";
 import LoadingAnimation from "../components/LoadingAnimation";
 import { setTitle } from "../assorted/setTitle";
 import Match from "./exerciseTypes/match/Match";
 import strings from "../i18n/definitions";
-import ImproveTranslationButtons from "./ImproveTranslationButtons";
+import FeedbackDisplay from "./bottomActions/FeedbackDisplay";
 
 let BOOKMARKS_TO_PRACTICE = 10;
 
@@ -43,7 +42,8 @@ export default function Exercises({ api, articleID }) {
   const [currentExerciseType, setCurrentExerciseType] = useState(null);
   const [isCorrect, setIsCorrect] = useState(false);
   const [showFeedbackButtons, setShowFeedbackButtons] = useState(false);
-  const [showWordSelector, setShowWordSelector] = useState(false);
+  const [showNewTranslationPrompt, setShowNewTranslationPrompt] =
+    useState(false);
 
   useEffect(() => {
     if (exerciseSession.length === 0) {
@@ -257,13 +257,13 @@ export default function Exercises({ api, articleID }) {
   }
 
   function toggleShow() {
-    if (showWordSelector) setShowWordSelector(false);
+    if (showNewTranslationPrompt) setShowNewTranslationPrompt(false);
     setShowFeedbackButtons(!showFeedbackButtons);
   }
 
   function toggleShowImproveTranslation() {
     if (showFeedbackButtons) setShowFeedbackButtons(false);
-    setShowWordSelector(!showWordSelector);
+    setShowNewTranslationPrompt(!showNewTranslationPrompt);
   }
 
   let wordSourceText = articleInfo ? (
@@ -294,19 +294,15 @@ export default function Exercises({ api, articleID }) {
           toggleShowImproveTranslation={toggleShowImproveTranslation}
         />
       </s.ExForm>
-      <FeedbackButtons
-        show={showFeedbackButtons}
-        setShow={setShowFeedbackButtons}
-        feedbackFunction={stopShowingThisFeedback}
-        currentExerciseType={currentExerciseType}
-        currentBookmarksToStudy={currentBookmarksToStudy}
-      />
-      <ImproveTranslationButtons
-        show={showWordSelector}
-        setShow={setShowWordSelector}
+      <FeedbackDisplay
+        showFeedbackButtons={showFeedbackButtons}
+        setShowFeedbackButtons={setShowFeedbackButtons}
+        showNewTranslationPrompt={showNewTranslationPrompt}
+        setShowNewTranslationPrompt={setShowNewTranslationPrompt}
         currentExerciseType={currentExerciseType}
         currentBookmarksToStudy={currentBookmarksToStudy}
         adjustCurrentTranslation={adjustCurrentTranslation}
+        feedbackFunction={stopShowingThisFeedback}
       />
     </s.ExercisesColumn>
   );
